--- src/raytracer-engine/types.d.ts
@@ -1,4 +1,0 @@
-declare module '*.wgsl' {
-  const shader: 'string';
-  export default shader;
-}
