--- src/utils/shader.d.ts
@@ -1,1 +1,0 @@
-declare module "*.wgsl"
